+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 1.1.0 - November 15, 2024
 
 - ✨ Added new ``wpkirk_md` helper function to parse the backticks in the string.
-- Improved  the all helpers functions.
+- ✨ Improved the all helpers functions.
 - Removed the useless functions.
 
 ## 1.0.0 - October 15, 2024